Fix aligned bug and move binding constants out of constants.java
