+++ ThriftyTest/src/main/java/frc/robot/Constants.java
@@ -9,11 +9,9 @@ import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.Distance;
 import edu.wpi.first.units.measure.Time;
-import edu.wpi.first.wpilibj.PS5Controller.Button;
 import frc.robot.subsystems.elevator.ElevatorState;
 
 public class Constants {
-
   public static class IDConstants {
     public static final int candle1 = 5;
     public static final int candle2 = 6;
@@ -25,114 +23,6 @@ public class Constants {
 
   public static class RobotConstants {
     public static final Time globalCanTimeout = Milliseconds.of(20); // 20 milliseconds
-  }
-
-  public static class ButtonBindingConstants {
-    public static enum DriverChoice {
-      DRAGONREINS, BACKUP;
-    }
-    public static enum ButtonBoardChoice {
-      PS5, KEYBOARD;
-    }
-
-    public static final DriverChoice driverChoice = DriverChoice.DRAGONREINS;
-    public static final ButtonBoardChoice buttonBoardChoice = ButtonBoardChoice.PS5;
-
-    public static final String dragonReinsName = "spark";
-    public static final String driverBackupName = "inter";
-
-    public static final String ps5Name = "dual";
-
-    public static final int driverPort = 0;
-    public static final int operatorPort = 1;
-
-    public static class Driver {
-      public static final int xAxis = 1;
-      public static final int yAxis = 0;
-      public static final int rotAxis = 3;
-
-      public static final boolean flipX = false;
-      public static final boolean flipY = true;
-      public static final boolean flipRot = false;
-
-      public static final int resetHeading = 1;
-      public static final int processor = 2;
-
-      public static final double deadband = 0.01;
-    }
-
-    public static class Operator {
-      public static final int L1 = 180; // POV
-      public static final int L2 = 270; // POV
-      public static final int L3 = 90; // POV
-      public static final int L4 = 0; // POV
-
-      public static final int secondaryL1 = 14;
-
-      public static final int ejectCoral = Button.kL2.value;
-
-      public static final int leftReef = Button.kSquare.value;
-      public static final int rightReef = Button.kCircle.value;
-
-      public static final int lowAlgae = Button.kCross.value;
-      public static final int highAlgae = Button.kTriangle.value;
-      public static final int ground = 180; // POV
-      public static final int processor = 90; // POV
-      public static final int highGround = 270; // POV
-      public static final int net = 0; // POV
-      public static final int algaeModeButton = Button.kR2.value; // R2
-
-      public static final int autoProcessor = Button.kR1.value;
-
-      public static final int climbUp = Button.kCreate.value;
-      public static final int climb = Button.kOptions.value;
-
-      public static final int stow = Button.kPS.value;
-
-      public static final int intake = Button.kL1.value; // LB
-
-      public static final int zeroElevator = 15; // old safety mode button (little bar below PS
-                                                 // button)
-
-      public static final int rightFunnel = 11;
-      public static final int leftFunnel = 12;
-
-
-    }
-
-    public static class ButtonBoardKeyboard {
-      // WHEN SAFETY ON - AUTOMATION BASED
-      public static final int L1 = 1;
-      public static final int L2 = 2;
-      public static final int L3 = 3;
-      public static final int L4 = 4;
-
-      public static final int A = 5;
-      public static final int B = 6;
-      public static final int C = 7;
-      public static final int D = 8;
-      public static final int E = 9;
-      public static final int F = 10;
-      public static final int G = 11;
-      public static final int H = 12;
-      public static final int I = 13;
-      public static final int J = 14;
-      public static final int K = 15;
-      public static final int L = 16;
-
-      public static final int lowAlgae = 17;
-      public static final int highAlgae = 18;
-      public static final int groundAlgae = 19;
-      public static final int processor = 20;
-      public static final int net = 21;
-
-      public static final int leftIntake = 22;
-      public static final int rightIntake = 23;
-
-      public static final int climb = 24;
-
-      public static final int cancelAuto = 25;
-    }
   }
 
   public static class FieldConstants {
@@ -284,10 +174,6 @@ public class Constants {
     private ScoringLocationsCenter(Pose2d pose) {
       value = pose;
     }
-  }
-
-  public enum ReefClipLocations {
-    LEFT, RIGHT;
   }
 
   public static class LedConstants {
