+++ ThriftyTest/src/main/java/frc/robot/binding/DriveBindings.java
@@ -3,20 +3,19 @@ package frc.robot.binding;
 import java.util.function.DoubleSupplier;
 import edu.wpi.first.wpilibj2.command.button.CommandPS5Controller;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.Constants.ButtonBindingConstants;
-import frc.robot.Constants.ButtonBindingConstants.Driver;
 import frc.robot.Constants.FieldConstants;
+import frc.robot.binding.BindingConstants.Driver;
 import frc.robot.driveassist.APTarget;
 import frc.robot.superstructure.Superstructure;
 import frc.robot.superstructure.states.Align;
-import frc.robot.superstructure.states.HeadingReset;
 import frc.robot.superstructure.states.DeferredAlign;
-import frc.robot.superstructure.states.TeleopDrive;
 import frc.robot.superstructure.states.DeferredAlign.AlignLocation;
+import frc.robot.superstructure.states.HeadingReset;
+import frc.robot.superstructure.states.TeleopDrive;
 
 public class DriveBindings implements Binder {
   private final CommandPS5Controller m_controller =
-      new CommandPS5Controller(ButtonBindingConstants.driverPort);
+      new CommandPS5Controller(BindingConstants.driverPort);
 
   private final DoubleSupplier m_x =
       () -> m_controller.getRawAxis(Driver.xAxis) * (Driver.flipX ? -1.0 : 1.0);;
